requirements.txt
celery==5.2.7
Django==4.1.5
django-appconf==1.0.5
django-compressor==4.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
eventlet==0.33.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
psycopg2==2.9.5
python-dotenv==0.21.1